add LogOfChanges

+++ src/main/java/org/example/dto/StockItemReq.java
@@ -2,6 +2,6 @@ package org.example.dto;
 
 import org.jetbrains.annotations.NotNull;
 
-public record StockItemReq (@NotNull String productId,@NotNull int warehouseLocation,@NotNull int quantity){
+public record StockItemReq (@NotNull String productId,@NotNull int warehouseLocation,@NotNull long quantity){
 
 }
